src/whispero/app.py
from __future__ import annotations

import os
import platform
import signal
import sys
import threading
from pathlib import Path

import requests
from pynput import keyboard

from .audio import RecorderState, start_recording, stop_recording
from .clipboard import paste_text
from .config import load_config
from .dictionary import load_dictionary, open_dictionary
from .sounds import play_sound
from .transcribe import transcribe

signal.signal(signal.SIGINT, lambda *_: (print("\n👋 Bye!"), os._exit(0)))

config = load_config()
state = RecorderState()


KEY_MAP = {
    "win": keyboard.Key.cmd,
    "cmd": keyboard.Key.cmd,
    "cmd_r": keyboard.Key.cmd_r,
    "ctrl": keyboard.Key.ctrl_l,
    "ctrl_r": keyboard.Key.ctrl_r,
    "shift": keyboard.Key.shift,
    "shift_r": keyboard.Key.shift_r,
    "alt": keyboard.Key.alt,
    "alt_r": keyboard.Key.alt_r,
}


def _bundle_dir() -> Path:
    if getattr(sys, "frozen", False):
        return Path(sys._MEIPASS)
    return Path(__file__).resolve().parents[2]


def _sounds_dir() -> Path:
    root = _bundle_dir()
    if getattr(sys, "frozen", False):
        return root / "sounds"
    return root / "assets" / "sounds"


def _dictionary_seed_path() -> Path:
    root = _bundle_dir()
    return root / "dictionary.txt"


def _play_sound(name: str) -> None:
    play_sound(name=name, sounds_enabled=bool(config.get("sounds", True)), sounds_dir=_sounds_dir())


def get_trigger_keys() -> set:
    """Get trigger keys from config based on platform."""
    is_mac = platform.system() == "Darwin"
    key_names = config["hotkey"].get("mac" if is_mac else "windows", ["cmd", "ctrl"])
    keys = set()
    for name in key_names:
        if name.lower() in KEY_MAP:
            keys.add(KEY_MAP[name.lower()])
        else:
            print(f"  ⚠️  Unknown key: {name}")
    return keys


def on_hotkey_press() -> None:
    start_recording(state, _play_sound)


def on_hotkey_release() -> None:
    audio_buf = stop_recording(state, _play_sound)
    if audio_buf is None:
        return

    def do_transcribe() -> None:
        prompt = load_dictionary(seed_path=_dictionary_seed_path())
        text = transcribe(audio_buf=audio_buf, config=config, prompt=prompt)
        if text:
            print(f"  📝 \"{text}\"")
            paste_text(text)
            print("  ✅ Pasted!")
        else:
            print("  ⚠️  No transcription returned")

    threading.Thread(target=do_transcribe, daemon=True).start()


def create_tray_icon():
    """Create and run the system tray icon."""
    try:
        import pystray
        from PIL import Image, ImageDraw
    except ImportError:
        print("  ⚠️  pystray/Pillow not installed, running without tray icon")
        return None

    def make_icon():
        # Try to load the 😮 icon from bundled or project icons
        icon_paths = [
            _bundle_dir() / "icons" / "icon_128.png",
            _bundle_dir() / "icons" / "icon.png",
            Path(__file__).resolve().parents[2] / "icons" / "icon_128.png",
            Path(__file__).resolve().parents[2] / "icons" / "icon.png",
        ]
        for p in icon_paths:
            if p.exists():
                return Image.open(p).resize((64, 64), Image.LANCZOS)
        # Fallback: 😮 face (yellow circle, two eyes, open mouth)
        img = Image.new("RGBA", (64, 64), (0, 0, 0, 0))
        draw = ImageDraw.Draw(img)
        draw.ellipse([4, 4, 60, 60], fill="#FFD93D", outline="#2D2D2D", width=3)
        draw.ellipse([20, 22, 28, 30], fill="#2D2D2D")  # left eye
        draw.ellipse([36, 22, 44, 30], fill="#2D2D2D")  # right eye
        draw.ellipse([26, 38, 38, 50], fill="#2D2D2D")  # open mouth
        return img

    def on_toggle(icon, item):
        state.enabled = not state.enabled
        status = "Enabled" if state.enabled else "Disabled"
        print(f"  🔄 Dictation {status}")

    def on_quit(icon, item):
        icon.stop()
        os._exit(0)

    def on_edit_dict(icon, item):
        open_dictionary()

    if platform.system() == "Darwin":
        hotkey_label = "Hold ⌃⌘ to dictate"
    else:
        hotkey_label = "Hold Win+Ctrl to dictate"

    menu = pystray.Menu(
        pystray.MenuItem(hotkey_label, None, enabled=False),
        pystray.Menu.SEPARATOR,
        pystray.MenuItem(lambda item: "✓ Enabled" if state.enabled else "  Disabled", on_toggle),
        pystray.MenuItem("Edit Dictionary", on_edit_dict),
        pystray.Menu.SEPARATOR,
        pystray.MenuItem("Quit", on_quit),
    )

    icon = pystray.Icon("WhisperO", make_icon(), "WhisperO", menu)
    return icon


def main() -> None:
    backend = config.get("backend", "local")

    if backend == "local":
        try:
            from .transcribe import get_model, is_model_cached
            model_name = config.get("model", "large-v3")
            print(f"😮 WhisperO (local, model: {model_name})")
            if not is_model_cached(model_name):
                print("  ⏳ Downloading model (this may take a few minutes)...")
            else:
                print("  ⏳ Loading model...")
            get_model(model_name)
            print("  ✓ Model ready")
        except (ImportError, RuntimeError):
            print("  ⚠️  faster-whisper not available, falling back to server mode")
            backend = "server"
            config["backend"] = "server"
    if backend == "server":
        print(f"🎤 WhisperO (server: {config['server']})")
        try:
            response = requests.get(f"{config['server']}/health", timeout=5)
            if response.json().get("status") == "ok":
                print("  ✓ Server is healthy")
            else:
                print("  ⚠️  Unexpected server response")
        except Exception:
            print("  ❌ Cannot reach server, will retry on each recording")

    trigger_keys = get_trigger_keys()
    keys_held = set()
    recording_active = False

    is_mac = platform.system() == "Darwin"
    if is_mac:
        key_names = config["hotkey"].get("mac", ["cmd", "ctrl"])
        print(f"🎹 Hotkey: hold [{' + '.join(k.title() for k in key_names)}] to record")
    else:
        key_names = config["hotkey"].get("windows", ["win", "ctrl"])
        print(f"🎹 Hotkey: hold [{' + '.join(k.title() for k in key_names)}] to record")
    print("🔇 Press Ctrl+C to quit\n")

    def on_press(key):
        nonlocal recording_active
        keys_held.add(key)
        if trigger_keys.issubset(keys_held) and not recording_active:
            recording_active = True
            on_hotkey_press()

    def on_release(key):
        nonlocal recording_active
        if key in trigger_keys and recording_active:
            recording_active = False
            on_hotkey_release()
        keys_held.discard(key)

    listener = keyboard.Listener(on_press=on_press, on_release=on_release)
    listener.daemon = True
    listener.start()

    tray = create_tray_icon()
    if tray:
        tray.run()
    else:
        try:
            listener.join()
        except KeyboardInterrupt:
            print("\n👋 Bye!")


if __name__ == "__main__":
    main()
